Since we upgraded to Java 7, take the opportunity for updating code that can take advantage of new API. Most of the time the changes are related to syntatic sugar, without change in functionalities. But in a few cases there is changes in the way exceptions are handled (e.g. try-with-resources).

+++ geoapi-conformance/src/main/java/org/opengis/test/referencing/AffineTransformTest.java
@@ -222,7 +222,7 @@ public strictfp class AffineTransformTest extends TransformTestCase {
         reference.transform(source, 0, target, 0, coordinates.length/2);
         verifyTransform(source, target);
         for (int i=0; i<coordinates.length; i++) {
-            assertEquals("Source array should be unmodified.", (double) coordinates[i], source[i], 0.0);
+            assertEquals("Source array should be unmodified.", coordinates[i], source[i], 0.0);
         }
     }
